fix: Move new BlogPage to standard folder
