+++ src/redesign/components/HomeBlogPreview.jsx
@@ -330,7 +330,7 @@ export function FeaturedBlogPreview({ blogs, width, imageHeight }) {
   const currentBlog = blogs[currentBlogIndex] || {};
   const imageUrl = currentBlog.image
     ? require("../images/posts/" + currentBlog.image)
-    : require("../images/placeholder.png");
+    : require("../../images/placeholder.png");
   const countryId = useCountryId();
   const link = `/${countryId}/research/${currentBlog.slug}`;
   return (
@@ -391,7 +391,7 @@ export function MediumBlogPreview({ blog, minHeight }) {
   const countryId = useCountryId();
   const imageUrl = blog.image
     ? require("../images/posts/" + blog.image)
-    : require("../images/placeholder.png");
+    : require("../../images/placeholder.png");
   const slug = blog.filename.split(".")[0];
   const link = `/${countryId}/research/${slug}`;
   return (
